Add permutations over same-sign leptons and b-jets
Basically, we have 2 same-sign leptons and 2 b-jets. If we denote the leptons by L1 and L2, and b-jet by B1 and B2, then the possible permutations are:
 - L1, L2, B1, B2;
 - L2, L1, B1, B2;
 - L1, L2, B2, B1;
 - L2, L1, B2, B1.
This is exactly what the code does by wrapping a class (IndexWrapper) around the leptons and jets which keeps track of current index representation of the particles. Th
e user can only choose the representation of particular permutation but not modify the MeasuredEvent instance inside integrate() functions.