Fix hadronic recoil transfer function
Previously we missed the effect that the true reconstructed and measured reconstructed b-jet energies (and therefore both 4-momenta) are different, and this should be included in the calculation of the hadronic recoil difference. For further information, see AN2013_313_v7 (eq. 14).

Also added the values returned by the transfer functions to debug plotting facility.