Limit the number of calls in Monte Carlo integration
Tested w/ tth and ttz samples that passed the generator level selection.
Basically, MXMC performs an initial search for non-zero probability
value. In our previous configurations we used 1M calls, but this is not
a viable option b/c most of the events that hit the 1M mark are lost
anyways. The optimal limiting number for this initial stage has been
found by plotting the number of calls it takes to reach a non-zero
probability. Found out that there's a gap from 1k to 1M, so it's
reasonable to limit the initial number of calls to 1k.

Also disable debug plots by default, and limit the number of events
processed per job to 1.