Create a set of jobs for testing various MEM settings
Added a meta-script, which generates a bunch of jobs with different configurations, while re-using the existing code as much as possible. The jobs can be generated with the following studies in mind:
1) various Higgs widths;
2) various initial number of calls in Markov Chain integrator (MXMC);
3) various number of integration calls (any integrator);
4) various clamping scenarios (gen level: all clamped, none clamped, z1 clamped, all but z1 clamped; reco level none clamped).

Have only tested the configuration generation for cases 1) and 4), which passed successfully. Haven't run any jobs from these studies, tho.

The motivation to create such automation tool is that it was kinda painful to run the MEM jobs manually every time I needed to generate the jobs for a range of settings (e.g. for a range of Higgs widths).

The resulting ROC curves can be aggregated via scripts/plotROCcurve.py by passing the list of subdirectories generated by python/createJobPool_3l1tau.py.