Proper event recording in debug tree
Pervious implementation didn't take into account the possibility that the MEM scores are evaluated for multiple combinations of b-jets, which varies case-by-case. Each TDirectory contained multiple cycles of TTrees, which corresponded to the b-jet combinations, but it was impossible to figure out which combination corresponded to which cycle. Refactored the logic and removed confusing debugRange variable.

Also added METx and METy differences to the debug TTree.