Default value for MXMC counters in the output TTree
Previously didn't consider the case in which the number of integration dimensions is zero (although the chosen integrator was Markov Chain). Also, if there was an error during the MEM score evaluation, the MXMC counter can be garbage. Also encountered a type conversion mismatch which wasn't a bug in itself (b/c long long covers unsigned type) but still... Forgot to comment some code in the previous commit.