Continuous resubmission of jobs
Limit the number of jobs and continuously submit the jobs until either the limit is reached or if there's nothing to submit anymore. Also, if for some reason the target failed, resubmit until reasonable number of retries has been reached. If the number of retries is exceeded, the make command will fail since not all targets are present in order to proceed.