Fuzzy limit of the number of jobs in the job pool
Associate a GUID to each job pool; each worker periodically checks the number of running jobs in the squeue that has the GUID as the comment, so that we can treat the whole pool as a homogenous unit, the number of jobs of which we want to limit. If the number of such jobs exceeds the pre-set limit, the job submission comes to a halt until the number of running jobs drops below to the said limit. This continues until all jobs (including resubmissions) are exhausted. The squeue command is kept in another thread so that the updating command isn't issued every time we loop over the list of jobs (which is very fast). The countereffect here is that the limit becomes more fuzzy (i.e. slightly above the specified limit) since multiple workers are submitting the jobs even though the number of actual jobs has already exceeded the specified limit because the numbers haven't been updated yet. It's not a problem in itself, fortunately.