Fix the issue w/ temporary job files
The problem was that the JobCreator class was initialized wrongly, which caused the jobs to write to a local directory. Since the temporary output directory and filenames coincided, the files were constantly overwritten in parallel and hence they were wrong. Some jobs even stalled on the nodes (and they still do...).

Also removed 2.5k nof integration points from nof-calls study.