Restructure the project
Discovered a way to place source files into different folders without breaking the build. The reason I did this is because there will be a bunch of files in the repository generated by MadGraph 5. The restructured verison also gives a better overview of the project.

Needs a complete re-build of the whole workspace, though.